pyproject.toml
[project]
name = "taltos"
version = "0.1.0"
description = "Multi-model MCP server exposing Gemini and Grok as tools to Claude Code"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Mátyás Rada" },
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.0",
    "openai>=2.32.0",
]

[project.scripts]
taltos = "taltos.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/taltos"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (undefined names, unused imports)
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear (common pitfalls)
    "UP",  # pyupgrade (modern Python syntax)
]

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.11",
]
